add `lib` directory, add `shorter/` alias
