+++ deno.lock
@@ -107,6 +107,9 @@
     "https://deno.land/x/discord_api_types@0.37.58/utils/internals.ts": "cb70895ba89f7947c38f7fa447b0190cb14b5585be323414cda53d2ccb19b16c",
     "https://deno.land/x/discord_api_types@0.37.58/utils/v10.ts": "056bd036f8c65365ff28eb63ec6897811d51921cca6d068392dd1ca5b397ae62",
     "https://deno.land/x/discord_api_types@0.37.58/v10.ts": "f3f23492c59e77859aba5b34431edf3668c37f722d7f70c2e1ef7ba4bcda3010",
+    "https://deno.land/x/durationjs@v4.1.0/mod.ts": "a7f9d3f38823caa9965f3e61f74a75fb09be548cca2517bd91be7ce15e4b35f1",
+    "https://deno.land/x/durationjs@v4.1.0/src/duration.ts": "e0c9a09311c22a6dafb7429081802c85a7920176213215315cbf587c5ea7a5ac",
+    "https://deno.land/x/durationjs@v4.1.0/src/in_words.ts": "1839490e2cb17fec2871bf87d6c7f7435c68416f9360d9f140777936d4680137",
     "https://deno.land/x/github_api_types@2023-05-17-05-41/mod.ts": "dc3a5cd3176c78085b49601e9c3fccac24809b037929230293255edabafbd0bb",
     "https://esm.sh/fast-json-patch@3.1.1": "fd59bab1fabdb3e7e2ce8204aa92113dc451708797021fff6f96b8ff265faedf",
     "https://esm.sh/tweetnacl@1.0.3": "fe5086fc2857018dc6118351ec0790016454d787af13301d65331c9a08eabb6a",
